requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
aiohttp

# ============================================================================
# PyTorch с CUDA 12.4 (установить ПЕРВЫМ!)
# ============================================================================
# ВАЖНО: Установите PyTorch с CUDA перед остальными зависимостями:
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128

# ============================================================================
# F5-TTS и аудио обработка
# ============================================================================
f5-tts>=0.1.0
soundfile>=0.12.1
librosa>=0.10.1
pydub>=0.25.1
numpy>=1.24.0,<2.0.0
faster-whisper>=0.10.0
ruaccent>=2.0.0

# redis - removed, using database instead
# aioredis - removed, using database instead
